Fix gcc warnings about unused vars.
It seems to me that at least in Ay_Cpu that data might be the same as data2 in
this particular case, and so it might be safe to remove data2. I can't prove it
for Kss_Cpu what with all the switch usage, but since we're not using the value
anyways we might have a bug going on by using data instead of data2.


git-svn-id: http://game-music-emu.googlecode.com/svn@37 ed62ecc2-ceed-11dd-a8b6-f5aca7b582bb